refactor: fix 3 cross-layer dependency violations
1. agentic_response.py: cli/ (L5) → llm/ (L2). LLM providers no longer
   import from CLI layer. cli/agentic_response.py kept as re-export stub.
2. MODEL_PRICING re-export removed from config.py (L1→L2). No callers
   used core.config.MODEL_PRICING — dead backward-compat import.
3. RightsRiskResult/RightsStatus/LicenseInfo: verification/ (L3) → state.py (L1).
   State layer no longer depends on verification layer. rights_risk.py
   re-exports from core.state for backward compatibility.
