+++ core/state.py
@@ -6,11 +6,40 @@ Domain layer: pure data models with no infrastructure dependencies.
 from __future__ import annotations
 
 import operator
+from enum import StrEnum
 from typing import Annotated, Any, Literal, TypedDict
 
 from pydantic import BaseModel, Field, model_validator
 
-from core.verification.rights_risk import RightsRiskResult
+
+class RightsStatus(StrEnum):
+    """IP rights clearance status."""
+
+    CLEAR = "clear"
+    NEGOTIABLE = "negotiable"
+    RESTRICTED = "restricted"
+    EXPIRED = "expired"
+    UNKNOWN = "unknown"
+
+
+class LicenseInfo(BaseModel):
+    """License details for an IP holder."""
+
+    holder: str
+    status: RightsStatus
+    expiry_year: int | None = None
+    territories: list[str] = Field(default_factory=list)
+    exclusivity: bool = False
+
+
+class RightsRiskResult(BaseModel):
+    """Result of IP rights/license risk assessment."""
+
+    status: RightsStatus
+    risk_score: int = Field(ge=0, le=100)
+    license_info: LicenseInfo
+    concerns: list[str] = Field(default_factory=list)
+    recommendation: str = ""
 
 _ITERATION_HISTORY_MAX = 10
